+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Chapters
 - Subplots and chapters do not repeat plot scripts
 - Scriptable.no_scripts is deprecated
+- Refactoring/removing unused methods
 
 ## 3.3.0 - September 1, 2024
 - Node#take